+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+## [2.2.0] - 2024-02-22
+### Breaking
+- onClick now becomes onTapped
+- TouchFeedBack on IOS now more inline with the native look and feel
+- Added posiblity to add custom effects
+- Removed BetterInkwell widget
+
+# Changelog
 ## [2.1.5] - 2024-02-12
 ### Changed
 -  Fixed missing material in the TouchFeedback widget